refactor: userCounter 컴포넌트 분리
